+++ CHANGELOG.md
@@ -32,7 +32,7 @@
   delimiter in the output.
 - Pressing the Action keycap on the virtual keyboard with the shift key pressed
   now inserts a new line (similar to what shift+enter does on a physical keyboard).
-  
+- Render `\displaystyle` and `\textstyle` to MathML
 
 ## 0.95.5 (2023-08-18)